Restructure packages for future work (#27)
## Describe your changes
I restructure the code so that future work makes a bit more sense.

## Checklist before requesting a review
- [x] The code runs successfully.

```
(poetry-python-project-py3.11) michaelp@MacBook-Air-3 supreme-court-ml-predictions % make format
isort "supreme_court_predictions"/ test/
Skipped 1 files
black "supreme_court_predictions"/ test/
All done! ✨ 🍰 ✨
16 files left unchanged.
(poetry-python-project-py3.11) michaelp@MacBook-Air-3 supreme-court-ml-predictions % make lint
ruff "supreme_court_predictions"/ test/
(poetry-python-project-py3.11) michaelp@MacBook-Air-3 supreme-court-ml-predictions % make test
pytest -vs test/
=================================================================================== test session starts ===================================================================================
platform darwin -- Python 3.11.3, pytest-7.3.1, pluggy-1.0.0 -- /Users/michaelp/Documents/GitHub/supreme-court-ml-predictions/.venv/bin/python
cachedir: .pytest_cache
rootdir: /Users/michaelp/Documents/GitHub/supreme-court-ml-predictions
collected 1 item                                                                                                                                                                          

test/test_util.py::test_get_full_pathway PASSED

==================================================================================== 1 passed in 0.00s ====================================================================================
(poetry-python-project-py3.11) michaelp@MacBook-Air-3 supreme-court-ml-predictions % 
```
- [x] I ran `make lint` and have made the bulk of changes that it has
requested.
