+++ supreme_court_predictions/__main__.py
@@ -4,7 +4,7 @@ This file serves as the entry point for the supreme_court_predictions module.
 
 import argparse
 
-from supreme_court_predictions.api.convokit.client import get_data
+from supreme_court_predictions.api.convokit import get_data
 from supreme_court_predictions.statistics.service import (
     clean_data,
     describe_data,
@@ -33,14 +33,6 @@ if __name__ == "__main__":
     )
 
     parser.add_argument(
-        "--tokenize-data",
-        help="Generate tokenizations for Convokit",
-        type=bool,
-        default=False,
-        action=argparse.BooleanOptionalAction,
-    )
-
-    parser.add_argument(
         "--describe-data",
         help="Generate descriptive statistics for Convokit",
         type=bool,
@@ -55,8 +47,6 @@ if __name__ == "__main__":
 
     if args.clean_data:
         clean_data()
-
-    if args.tokenize_data:
         tokenize_data()
 
     if args.describe_data:
