+++ supreme_court_predictions/data/models/README.md
@@ -1,2 +1,2 @@
 ## Models Data Folder
-This folder contains all cleaned Models data. The models data will not be saved in GitHub due to the storage demand.
+This folder contains all cleaned models' data. The models data will not be saved in GitHub due to the storage demand.
