+++ supreme_court_predictions/data/processed/README.md
@@ -1,0 +1,2 @@
+## Processed Data Folder
+This folder contains all cleaned processed data. The processed data will not be saved in GitHub due to the storage demand.
