+++ supreme_court_predictions/data/statistics/README.md
@@ -1,2 +1,2 @@
 ## Statistics Data Folder
-This folder contains all summary statistics data.
+This folder contains all summary statistics data. The summary statistics data will not be saved in GitHub due to the storage demand.
