+++ supreme_court_predictions/statistics/service.py
@@ -2,9 +2,9 @@
 This file works as the central point of interaction for the statistics package.
 """
 
-from supreme_court_predictions.statistics.datacleaner import DataCleaner
+from supreme_court_predictions.processing.datacleaner import DataCleaner
+from supreme_court_predictions.processing.tokenizer import Tokenizer
 from supreme_court_predictions.statistics.descriptives import Descriptives
-from supreme_court_predictions.statistics.tokenizer import Tokenizer
 
 
 def clean_data():
